backend/requirements.txt
Fabric
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
pymongo
redis
flask-cors